requirements.txt
openai>=1.40.0
chromadb>=0.5.5
tiktoken>=0.7.0
python-dotenv>=1.0.1
streamlit>=1.36.0
pydantic>=2.8.2
